Rename stuff to not have apollo in it
