+++ packages/graphql-server-core/src/graphqlOptions.ts
