+++ packages/graphql-server-core/src/index.ts
@@ -1,2 +1,2 @@
 export { runQuery, LogFunction, LogMessage, LogStep, LogAction } from './runQuery'
-export { default as ApolloOptions} from './apolloOptions'
+export { default as ApolloOptions} from './graphqlOptions'
